Extract MutedSyntaxWalker
